We introduce a new date picker which include month and year only in order to choose an expiry date for credit card (IBAInpuCreditCardExpiryDatePicker).
We also modify the way that default picker work. There is now an
IBAGenericPickerView, decoupled from IBAMultiplePickListInputProvider,
which adds the possibility of making your picker circular or not.

IBAPickerView inherits from this IBAGenericPickerView and is the
default picker.

Moreover, you can create your own custom picker. It just have to inherit
IBAGenericPickerView.

Also, your custom picker can be created with a specific image and/or
font by using an array of IBAPickListFormOption.

Check the IBAFormsShowcase sample project, it's up to date with
previous described changes.